Upgrade to snakeyaml/#337 for no-duplicate maps
This patch updates moultingyaml to use a patched snakeyaml
that had/has an unsavoury "feature" that allowed "duplicate" yaml maps.

In short without this patch, a yaml src like
```
foo:
  value: 1
foo:
  value: 2
```

will be valid, and will only have a "{foo:{value:2}}" map (no error).

With this change, you have the option to suppy configuration to force an exception to be thrown.

See the code change in https://bitbucket.org/asomov/snakeyaml/pull-requests/9 for more details
The snakeyaml version is "patched" as 1.18.i337. Once a committed version is pushed and released
for snakeyaml, this will turn into a PR to moultingyaml.
Issue https://bitbucket.org/asomov/snakeyaml/issues/337